Now outputs table of results by default and doesn't output object unless -PassThru is specified
I was getting sick of forgetting to pipe this module to Format-Table, and having to rerun it and either do that or capture the data for further formatting. So now outputting the data is only done when -PassThru is specified, and the data is displayed in a pre-formatted table (in the Info stream with the other logging) by default.